pyproject.toml
[tool.poetry]
name = "lang-analyze"
version = "0.1.0"
description = "figuring out what language some MDN docs are in"
authors = ["Peter Bengtsson <mail@peterbe.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.7"
guess_language-spirit = "^0.5.3"
pyquery = "^1.4.1"
bleach = "^3.1.1"
pyyaml = "^5.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
